third_party/WebKit/LayoutTests/fast/dom/HTMLObjectElement/object-as-frame-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS window.frames.length is 2
PASS window.frames[0].frameElement.name is "frame"
PASS window.frames[1].frameElement.name is "obj"
Loading 'resources/object-as-frame-subframe-new.html#updated-contents-of-iframe' in iframe...
Loading 'resources/object-as-frame-subframe-new.html#updated-contents-of-object' in object...
PASS window.frames[0].location.hash is "#updated-contents-of-iframe"
PASS window.frames[1].location.hash is "#updated-contents-of-object"
 
